멀티 윈도우 왜저러냐 ㅆㅃㄹ
ㅁㅇㄴㄻㄴㅇㄻㄴㅇㄹ

+++ KEngine2_0/190904_Circus/ComPlayer.cpp
@@ -242,7 +242,7 @@ void ComPlayer::update_jump()
 {
 	pAnimator->change_animation(L"Jump");
 	fJumpTime += KTimeManager::instance()->deltatime();
-	float T = cos(fJumpTime * 2.0f);
+	float T = cosf(fJumpTime * 2.0f);
 
 	kone()->moving_delta({ .0f, (T * fJumpPower) * -1.0f });
